+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- `s3ObjectVersionId` parameter to `FormsClient.GetFormSubmission()`
+- `formSubmissionMetaEdits`, `formSubmissionSchedulingBooking` and `taskCompletion` to `FormSubmissionMetadataResponse`
+- `lastEdit` to `FormSubmissionMetadata`
+- `approverEditability`, `requiresAllConditionallyShowOptionsPredicates`, `autocompleteAttributes`, `includeTimestampWatermark`, `layout` and `reverseGeocoding` to `FormElement`
+- nested `options` to `FormElementOption`
+
 ## [13.1.1] - 2026-08-26
 
 ### Added